team-service/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
requests
pika
PyJWT
django-cors-headers
django-storages
boto3
celery
django-celery-beat